Improve bulk_test.sh script (#5714)
* Preserve output of failed cases
* More testing environments
* Add slavonic fonts to testing suite

+++ tests/bulk_test.sh
@@ -15,11 +15,33 @@ FF_BIN_PROJECT=../build/bin/fontforge
 FF_PY_SCRIPT_LINE="f=fontforge.open(argv[1]); f.save(argv[2])"
 FF_OUTPUT_FILE="Output.sfd"
 
+# FF_PY_SCRIPT_LINE="f=fontforge.open(argv[1]); f.generate(argv[2], flags='no-FFTM-table')"
+# FF_OUTPUT_FILE="Output.ttf"
+
+# FF_PY_SCRIPT_LINE="\
+# f=fontforge.open(argv[1]);\
+# f.generateFeatureFile('Out.fea')"$'\n'"\
+# for l in f.gsub_lookups: f.removeLookup(l)"$'\n'"\
+# for l in f.gpos_lookups: f.removeLookup(l)"$'\n'"\
+# f.mergeFeature('Out.fea'); f.save(argv[2])"
+# FF_OUTPUT_FILE="Output.sfd"
+
+# # NOTE: Zeroize "CreationTime" and "ModificationTime" in SFD_DumpSplineFontMetadata()
+# FF_PY_SCRIPT_LINE="\
+# f=fontforge.open(argv[1]);\
+# f.cidConvertByCmap('/home/iorsh/devel/cmap-resources/Adobe-Identity-0/CMap/Identity-H')"$'\n'"\
+# f.generateFeatureFile('Out.fea')"$'\n'"\
+# for l in f.gsub_lookups: f.removeLookup(l)"$'\n'"\
+# for l in f.gpos_lookups: f.removeLookup(l)"$'\n'"\
+# f.mergeFeature('Out.fea'); f.save(argv[2])"
+# FF_OUTPUT_FILE="Output.sfd"
+
 SCRIPT_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )" &> /dev/null && pwd )
 FF_BIN_PROJECT=$SCRIPT_DIR/$FF_BIN_PROJECT
 
 cat $FONT_LIST |
 while read font_url; do
+    failure=0
     TMP_DIR_SYS=`mktemp -d`
     pushd $TMP_DIR_SYS > /dev/null
     if (wget -q --timeout=60 $font_url); then
@@ -36,17 +58,27 @@ while read font_url; do
         popd > /dev/null
 
         if [ ! -f $TMP_DIR_SYS/$FF_OUTPUT_FILE ] && [ ! -f $TMP_DIR_PROJ/$FF_OUTPUT_FILE ]; then
-            echo "NO OUTPUT: $font_url"
+            echo "\e[0;33mNO OUTPUT\e[0m: $font_url"
+            failure=1
         elif (cmp --quiet $TMP_DIR_SYS/$FF_OUTPUT_FILE $TMP_DIR_PROJ/$FF_OUTPUT_FILE); then
             echo "PASS: $font_url"
         else
             echo -e "\e[0;31mFAIL\e[0m: $font_url"
+            failure=1
         fi
 
-        rm -rf $TMP_DIR_PROJ
+        if [ $failure -eq 0 ]; then
+            rm -rf $TMP_DIR_PROJ
+        else
+            echo "    Project output: $TMP_DIR_PROJ"
+        fi
     else
         echo "SKIP: $font_url"
     fi
     popd > /dev/null
-    rm -rf $TMP_DIR_SYS
+    if [ $failure -eq 0 ]; then
+        rm -rf $TMP_DIR_SYS
+    else
+        echo "    System output: $TMP_DIR_SYS"
+    fi
 done
